Added Threading to help with performance.
The Group function now starts as soon as the Auth to devops is complete and the group function will check the user's membership in up to 25 groups at a time. LOTS of testing shows this is the sweet spot for larger orgs.  I may make the max parallel configurable at a later date.